Updated client and methods for autoupdates

+++ packages/parser/common.go
@@ -286,7 +286,7 @@ func InsertIntoBlockchain(transaction *model.DbTransaction, block *Block) error 
 		return err
 	}
 
-	err = autoupdate.TryUpdate(blockID)
+	err = autoupdate.TryUpdate(uint64(blockID))
 	if err != nil {
 		log.WithFields(log.Fields{"type": consts.AutoupdateError, "error": err}).Error("creating block")
 		return err
